fix: slider story to work
no throw errors
